Move each model to separate files
